refactor(session): extract session module to ozmux_session crate
Move session, cell, pane, and activity modules from daemon/core to the
ozmux_session crate. Replace OzmuxError/OzmuxResult in session code with
SessionError/SessionResult, and add From<SessionError> for OzmuxError in
core so existing ? conversions continue to work transparently.
